requirements.txt
numpy
h5py
Cython
Keras
scikit-learn
tensorflow-gpu
pyyaml
pylru
